Add support for race condition on start up of distributor.
The worker might be stuck without and indication of where the file
is located because the distributor might come up after the replicator job
message is sent to it.  This would leave the worker at the archive waiting
to be notified which distributor has the data it needs.   These modifications
will information the archive of that file's location, if this condition happens,
freeing the worker to contact the distributor.